Reorganize characters in their own module
