+++ src/characters/fighter/fighter.ts
@@ -1,0 +1,7 @@
+import { type Character } from "../types";
+
+export type Fighter = Character & {
+  weapon: string;
+  skill: number;
+  phrase: "Primero pego y luego pregunto";
+};
