+++ src/characters/king/king.ts
@@ -1,0 +1,6 @@
+import { type Character } from "../types";
+
+export type King = Character & {
+  yearsOfReign: number;
+  phrase: "Vais a morir todos";
+};
